+++ changelog.md
@@ -10,6 +10,7 @@
 * Highlight backquoted (``; `sym` ``) symbols in comments
 * Remove old find-definition mechanism in favor of the `,find` REPL
   command
+* Guess module name based on project structure in `fennel-proto-repl`
 
 ## 0.9.2 / 2025-03-13